plane table detection #7
