+++ dora_the_mug_finder_bringup/scripts/test_scene.py
@@ -1,18 +1,12 @@
 #!/usr/bin/env python3
 
-import csv
-import pickle
 from copy import deepcopy
-from random import randint
-from turtle import color
 
 import open3d as o3d
-import cv2
 import numpy as np
-import matplotlib.pyplot as plt
 from matplotlib import cm
-import math
 import os
+from plane_table_detection import PlaneDetection, PlaneTable
 
 
 view = {
@@ -36,32 +30,6 @@ view = {
 }
 
 
-class PlaneDetection():
-    def __init__(self, point_cloud):
-
-        self.point_cloud = point_cloud
-
-    def colorizeInliers(self, r,g,b):
-        self.inlier_cloud.paint_uniform_color([r,g,b]) # paints the plane in red
-
-    def segment(self, distance_threshold=0.03, ransac_n=5, num_iterations=50):
-
-        print('Starting plane detection')
-        self.plane_model, inlier_idxs = self.point_cloud.segment_plane(distance_threshold=distance_threshold, 
-                                                    ransac_n=ransac_n,
-                                                    num_iterations=num_iterations)
-        [self.a, self.b, self.c, self.d] = self.plane_model
-
-        self.inlier_cloud = self.point_cloud.select_by_index(inlier_idxs)
-        outlier_cloud = self.point_cloud.select_by_index(inlier_idxs, invert=True)
-        return outlier_cloud
-
-    def __str__(self):
-        text = 'Segmented plane from pc with ' + str(len(self.point_cloud.points)) + ' with ' + str(len(self.inlier_cloud.points)) + ' inliers. '
-        text += '\nPlane: ' + str(self.a) +  ' x + ' + str(self.b) + ' y + ' + str(self.c) + ' z + ' + str(self.d) + ' = 0' 
-        return text
-
-
 def main():
     
     # ------------------------------------------
@@ -69,8 +37,9 @@ def main():
     # ------------------------------------------
 
     files_path=f'{os.environ["DORA"]}'
+    
     # Scene dataset paths
-    filename = files_path + '/rgbd-scenes-v2/pc/14.ply'
+    filename = files_path + '/rgbd-scenes-v2/pc/04.ply'
     
     os.system('pcl_ply2pcd ' + filename + ' pcd_point_cloud.pcd')
     point_cloud_original = o3d.io.read_point_cloud('pcd_point_cloud.pcd')
@@ -78,6 +47,8 @@ def main():
     # ------------------------------------------
     # Execution
     # ------------------------------------------
+
+    #? find two planes: from the table and another
     point_cloud = deepcopy(point_cloud_original) 
     number_of_planes = 2
     colormap = cm.Set1(list(range(0,number_of_planes)))
@@ -98,28 +69,13 @@ def main():
         if len(planes) >= number_of_planes: # stop detection planes
             break
     
-    # diferença entre plane.c
-    distd = abs(planes[0].d) - abs(planes[1].d)
-    distb = abs(planes[0].b) - abs(planes[1].b)
-    print('dist_d=' + str(abs(distd)) + ' dist_b=' + str(distd))
-    
-    if abs(distd) > 0.3:
-        print('plano da mesa e plano chao')
-        if distd > 0:
-            plane = planes[1]
-        else:
-            plane = planes[0]     
-    else:
-        plane = planes[0]
-        print('plano mesa e outro')
-        
-    
+    #? find table plane
+    plane_table = PlaneTable(planes).planetable()
 
     # ------------------------------------------
     # Visualization
     # ------------------------------------------
-
-    entities = [plane.inlier_cloud]
+    entities = [plane_table.inlier_cloud]
     frame = o3d.geometry.TriangleMesh().create_coordinate_frame(size=3, origin=np.array([0., 0., 0.]))
     entities.append(point_cloud)
     entities.append(frame)
